@@ -612,6 +629,12 @@ elseif (TARGET_OS STREQUAL "linux")
         NvCodec/NvCodec
         /usr/local/cuda/include)
 
+    if (WIN32)
+      set(_CUDA_COMPILE_OPTIONS "-xcuda;--cuda-gpu-arch=sm_35;-std=gnu++20")
+    else()
+      set(_CUDA_COMPILE_OPTIONS "-xcuda;--cuda-gpu-arch=sm_35;-std=gnu++17")
+    endif()
+
     # これらのソースは CUDA としてコンパイルする
     set_source_files_properties(
         src/cuda/cuda_context_cuda.cpp
