@@ -23,6 +23,9 @@
 #include "hwenc_jetson/jetson_v4l2_capturer.h"
 #elif USE_NVCODEC_ENCODER
 #include "hwenc_nvcodec/nvcodec_v4l2_capturer.h"
+#elif USE_V4L2_ENCODER
+#include "hwenc_v4l2/libcamera_capturer.h"
+#include "hwenc_v4l2/v4l2_capturer.h"
 #endif
 #include "v4l2_video_capturer/v4l2_video_capturer.h"
 #else
@@ -159,6 +162,18 @@ int main(int argc, char* argv[]) {
     } else {
       return V4L2VideoCapturer::Create(std::move(v4l2_config));
     }
+#elif USE_V4L2_ENCODER
+    if (args.use_libcamera) {
+      LibcameraCapturerConfig libcamera_config = v4l2_config;
+      // use_libcamera_native == true でも、サイマルキャストの場合はネイティブフレームを出力しない
+      libcamera_config.native_frame_output =
+          args.use_libcamera_native && !(use_sora && args.sora_simulcast);
+      return LibcameraCapturer::Create(libcamera_config);
+    } else if (v4l2_config.use_native && !(use_sora && args.sora_simulcast)) {
+      return V4L2Capturer::Create(std::move(v4l2_config));
+    } else {
+      return V4L2VideoCapturer::Create(std::move(v4l2_config));
+    }
 #else
     return V4L2VideoCapturer::Create(std::move(v4l2_config));
 #endif
@@ -180,7 +195,6 @@ int main(int argc, char* argv[]) {
   rtcm_config.no_audio_device = args.no_audio_device;
 
   rtcm_config.fixed_resolution = args.fixed_resolution;
-  rtcm_config.show_me = args.show_me;
   rtcm_config.simulcast = args.sora_simulcast;
   rtcm_config.hardware_encoder_only = args.hw_mjpeg_decoder;
 
